Play 2.0-beta
I added a new formula as Play 2.0 is very different from 1.x, and that
1.x will still be supported going forward.  You should be able to have
both play and play2 installed at the same time.

There is an issue with the play script that causes it to not work if
you use a symlink with it.  This formula thus modifies the play script
to allow it to be installed successfully.

Signed-off-by: Ealden Esto E. Escanan <ealden@gmail.com>